Develop

+++ Front/src/Components/Home/Home.jsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
 import { useSelector } from "react-redux";
-import { useLocation /* , Link */ } from "react-router-dom";
+import { useLocation, Link } from "react-router-dom";
 import Card from "../Card/Card";
 import { firebaseApp } from "../Firebase/credentials";
 import { getDocs, collection, getFirestore } from "firebase/firestore";
@@ -8,6 +8,7 @@ import deleteRecordFS from "../../Utils/deleteRecordFS";
 import deleteImage from "../../Utils/deleteImage";
 import { IoMdCloudUpload } from "react-icons/io";
 import { AiFillTool } from "react-icons/ai";
+import Swal from "sweetalert2";
 import styles from "./Home.module.css";
 
 const firestore = getFirestore(firebaseApp);
@@ -16,6 +17,7 @@ export default function Home() {
   const location = useLocation().pathname === "/" ? "Home" : "Not Home";
 
   const [cardList, setCardList] = useState([]);
+  const [render, setRender] = useState(false); //Sólo sirve para forzar el re-render despues de eliminar un artículo
 
   const filter = useSelector((state) =>
     location === "Home" ? state.publicFilter : state.adminFilter
@@ -33,24 +35,52 @@ export default function Home() {
   }
 
   function handleDeleteCard(uid) {
-    deleteImage(uid);
-    deleteRecordFS(uid);
+    Swal.fire({
+      title: "¿Eliminar artículo?",
+      text: "Esta acción no es reversible",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Confirmar",
+      cancelButtonText: "Cancelar",
+      allowOutsideClick: false,
+    }).then((result) => {
+      if (result.isConfirmed) {
+        try {
+          deleteImage(uid);
+          deleteRecordFS(uid);
+          setRender(!render);
+          Swal.fire({
+            title: "Artículo eliminado correctamente",
+            // text: "Artículo eliminado correctamente",
+            icon: "success",
+          });
+        } catch (error) {
+          Swal.fire({
+            title: "Error al eliminar el artículo",
+            // text: "Error al eliminar el artículo",
+            icon: "error",
+          });
+        }
+      }
+    });
   }
 
   useEffect(() => {
     getDataFromFirestore();
-  }, []);
+  }, [render]);
 
   return (
     <div className={styles.homeContainer}>
       {location === "Not Home" && (
-        <div
-          className={styles.uploadContainer}
-          onClick={() => (window.location.href = "/admin/addItem")}
-        >
-          <IoMdCloudUpload className={styles.icon} />
-        </div>
+        <Link to="/admin/addItem" style={{ display: "contents" }}>
+          <div className={styles.uploadContainer}>
+            <IoMdCloudUpload className={styles.icon} />
+          </div>
+        </Link>
       )}
+
       {location === "Not Home" && (
         <div className={styles.toolContainer}>
           <AiFillTool className={styles.icon} />
